refactor(analytics): consolidate input validation into AnalyticsConverter
Devin Review (resolveThresholds の負値クランプ / tier1↔tier2 swap 抜け)
+ Gemini Review (両 usecase でロジック重複) を同時解決。

新ファイル: `src/application/domain/analytics/data/converter.ts`
- AnalyticsConverter.resolveThresholds
- AnalyticsConverter.clampHubBreadthThreshold
- AnalyticsConverter.clampDormantThresholdDays

完全版 (community 由来の負値クランプ + tier swap) を採用し、両 usecase
(analytics/usecase.ts と analytics/community/usecase.ts) から重複ロジックを
削除して新 converter を import するよう変更。analytics dashboard 側で
失われていた防御的処理が community detail と一致するようになった。

命名 / 配置: Standard Domain Pattern の `data/converter.ts` に揃える
(他ドメインの converter は `data/` 配下が標準)。

community usecase 専用の `clampLimit` (member-list 用) と analytics
dashboard 専用の `clampWindowDays` は移動なし — それぞれ 1 箇所でしか
使われない。

260 unit tests green、typecheck pass、動作変更ゼロ (community 経路は
完全互換、analytics dashboard 経路は防御強化)。
