+++ src/application/domain/analytics/usecase.ts
@@ -5,27 +5,15 @@ import {
   GqlAnalyticsDashboardPayload,
   GqlQueryAnalyticsDashboardArgs,
 } from "@/types/graphql";
-import AnalyticsCommunityService, {
-  DEFAULT_DORMANT_THRESHOLD_DAYS,
-  DEFAULT_HUB_BREADTH_THRESHOLD,
-  MAX_DORMANT_THRESHOLD_DAYS,
-  MAX_HUB_BREADTH_THRESHOLD,
-  MIN_DORMANT_THRESHOLD_DAYS,
-  MIN_HUB_BREADTH_THRESHOLD,
-} from "@/application/domain/analytics/community/service";
+import AnalyticsCommunityService from "@/application/domain/analytics/community/service";
 import AnalyticsPlatformService from "@/application/domain/analytics/platform/service";
-import {
-  DEFAULT_SEGMENT_THRESHOLDS,
-  MAX_MIN_MONTHS_IN,
-  MIN_MIN_MONTHS_IN,
-  SegmentThresholds,
-} from "@/application/domain/analytics/community/classifiers";
 import {
   computeDormantCount,
   computeStageCounts,
   computeTenureDistribution,
 } from "@/application/domain/analytics/community/aggregations";
 import AnalyticsPresenter from "@/application/domain/analytics/presenter";
+import AnalyticsConverter from "@/application/domain/analytics/data/converter";
 import { jstMonthStart, jstNextMonthStart } from "@/application/domain/report/util";
 import { asOfBounds } from "@/application/domain/analytics/community/bounds";
 
@@ -59,10 +47,14 @@ export default class AnalyticsUseCase {
     ctx: IContext,
   ): Promise<GqlAnalyticsDashboardPayload> {
     const asOf = input?.asOf ?? new Date();
-    const thresholds = resolveThresholds(input?.segmentThresholds);
+    const thresholds = AnalyticsConverter.resolveThresholds(input?.segmentThresholds);
     const windowDays = clampWindowDays(input?.windowDays);
-    const hubBreadthThreshold = clampHubBreadthThreshold(input?.hubBreadthThreshold);
-    const dormantThresholdDays = clampDormantThresholdDays(input?.dormantThresholdDays);
+    const hubBreadthThreshold = AnalyticsConverter.clampHubBreadthThreshold(
+      input?.hubBreadthThreshold,
+    );
+    const dormantThresholdDays = AnalyticsConverter.clampDormantThresholdDays(
+      input?.dormantThresholdDays,
+    );
 
     const monthStart = jstMonthStart(asOf);
     const platformUpperBound = asOfBounds(asOf).clampFuture(jstNextMonthStart(asOf));
@@ -114,33 +106,7 @@ export default class AnalyticsUseCase {
   }
 }
 
-function resolveThresholds(
-  input:
-    | { tier1?: number | null; tier2?: number | null; minMonthsIn?: number | null }
-    | null
-    | undefined,
-): SegmentThresholds {
-  if (!input) return DEFAULT_SEGMENT_THRESHOLDS;
-  const tier1 = input.tier1 ?? DEFAULT_SEGMENT_THRESHOLDS.tier1;
-  const tier2 = input.tier2 ?? DEFAULT_SEGMENT_THRESHOLDS.tier2;
-  const minMonthsIn = Math.min(
-    Math.max(input.minMonthsIn ?? DEFAULT_SEGMENT_THRESHOLDS.minMonthsIn, MIN_MIN_MONTHS_IN),
-    MAX_MIN_MONTHS_IN,
-  );
-  return { tier1, tier2, minMonthsIn };
-}
-
 function clampWindowDays(input: number | null | undefined): number {
   const n = input ?? DEFAULT_WINDOW_DAYS;
   return Math.min(Math.max(n, MIN_WINDOW_DAYS), MAX_WINDOW_DAYS);
-}
-
-function clampHubBreadthThreshold(input: number | null | undefined): number {
-  const n = input ?? DEFAULT_HUB_BREADTH_THRESHOLD;
-  return Math.min(Math.max(n, MIN_HUB_BREADTH_THRESHOLD), MAX_HUB_BREADTH_THRESHOLD);
-}
-
-function clampDormantThresholdDays(input: number | null | undefined): number {
-  const n = input ?? DEFAULT_DORMANT_THRESHOLD_DAYS;
-  return Math.min(Math.max(n, MIN_DORMANT_THRESHOLD_DAYS), MAX_DORMANT_THRESHOLD_DAYS);
 }
